move isPathAttribute to dataManager, use in link detector
