+++ src/services/htmlLinks.ts
@@ -8,12 +8,13 @@ import * as strings from '../utils/strings';
 import { URI as Uri } from 'vscode-uri';
 
 import { TokenType, DocumentContext, TextDocument, Range, DocumentLink } from '../htmlLanguageTypes';
+import { HTMLDataManager } from '../languageFacts/dataManager';
 
 function normalizeRef(url: string): string {
 	const first = url[0];
 	const last = url[url.length - 1];
 	if (first === last && (first === '\'' || first === '\"')) {
-		url = url.substr(1, url.length - 2);
+		url = url.substring(1, url.length - 1);
 	}
 	return url;
 }
@@ -85,63 +86,71 @@ function isValidURI(uri: string) {
 	}
 }
 
-export function findDocumentLinks(document: TextDocument, documentContext: DocumentContext): DocumentLink[] {
-	const newLinks: DocumentLink[] = [];
+export class HTMLDocumentLinks {
 
-	const scanner = createScanner(document.getText(), 0);
-	let token = scanner.scan();
-	let lastAttributeName: string | undefined = undefined;
-	let afterBase = false;
-	let base: string | undefined = void 0;
-	const idLocations: { [id: string]: number | undefined } = {};
-
-	while (token !== TokenType.EOS) {
-		switch (token) {
-			case TokenType.StartTag:
-				if (!base) {
-					const tagName = scanner.getTokenText().toLowerCase();
-					afterBase = tagName === 'base';
-				}
-				break;
-			case TokenType.AttributeName:
-				lastAttributeName = scanner.getTokenText().toLowerCase();
-				break;
-			case TokenType.AttributeValue:
-				if (lastAttributeName === 'src' || lastAttributeName === 'href') {
-					const attributeValue = scanner.getTokenText();
-					if (!afterBase) { // don't highlight the base link itself
-						const link = createLink(document, documentContext, attributeValue, scanner.getTokenOffset(), scanner.getTokenEnd(), base);
-						if (link) {
-							newLinks.push(link);
-						}
-					}
-					if (afterBase && typeof base === 'undefined') {
-						base = normalizeRef(attributeValue);
-						if (base && documentContext) {
-							base = documentContext.resolveReference(base, document.uri);
-						}
-					}
-					afterBase = false;
-					lastAttributeName = undefined;
-				} else if (lastAttributeName === 'id') {
-					const id = normalizeRef(scanner.getTokenText());
-					idLocations[id] = scanner.getTokenOffset();
-				}
-				break;
-		}
-		token = scanner.scan();
+	constructor(private dataManager: HTMLDataManager) {
 	}
-	// change local links with ids to actual positions
-	for (const link of newLinks) {
-		const localWithHash = document.uri + '#';
-		if (link.target && strings.startsWith(link.target, localWithHash)) {
-			const target = link.target.substr(localWithHash.length);
-			const offset = idLocations[target];
-			if (offset !== undefined) {
-				const pos = document.positionAt(offset);
-				link.target = `${localWithHash}${pos.line + 1},${pos.character + 1}`;
+
+	findDocumentLinks(document: TextDocument, documentContext: DocumentContext): DocumentLink[] {
+		const newLinks: DocumentLink[] = [];
+
+		const scanner = createScanner(document.getText(), 0);
+		let token = scanner.scan();
+		let lastAttributeName: string | undefined = undefined;
+		let lastTagName: string | undefined = undefined;
+		let afterBase = false;
+		let base: string | undefined = void 0;
+		const idLocations: { [id: string]: number | undefined } = {};
+
+		while (token !== TokenType.EOS) {
+			switch (token) {
+				case TokenType.StartTag:
+					lastTagName = scanner.getTokenText().toLowerCase();
+					if (!base) {
+						afterBase = lastTagName === 'base';
+					}
+					break;
+				case TokenType.AttributeName:
+					lastAttributeName = scanner.getTokenText().toLowerCase();
+					break;
+				case TokenType.AttributeValue:
+					if (lastTagName && lastAttributeName && this.dataManager.isPathAttribute(lastTagName, lastAttributeName)) {
+						const attributeValue = scanner.getTokenText();
+						if (!afterBase) { // don't highlight the base link itself
+							const link = createLink(document, documentContext, attributeValue, scanner.getTokenOffset(), scanner.getTokenEnd(), base);
+							if (link) {
+								newLinks.push(link);
+							}
+						}
+						if (afterBase && typeof base === 'undefined') {
+							base = normalizeRef(attributeValue);
+							if (base && documentContext) {
+								base = documentContext.resolveReference(base, document.uri);
+							}
+						}
+						afterBase = false;
+						lastAttributeName = undefined;
+					} else if (lastAttributeName === 'id') {
+						const id = normalizeRef(scanner.getTokenText());
+						idLocations[id] = scanner.getTokenOffset();
+					}
+					break;
+			}
+			token = scanner.scan();
+		}
+		// change local links with ids to actual positions
+		for (const link of newLinks) {
+			const localWithHash = document.uri + '#';
+			if (link.target && strings.startsWith(link.target, localWithHash)) {
+				const target = link.target.substring(localWithHash.length);
+				const offset = idLocations[target];
+				if (offset !== undefined) {
+					const pos = document.positionAt(offset);
+					link.target = `${localWithHash}${pos.line + 1},${pos.character + 1}`;
+				}
 			}
 		}
+		return newLinks;
 	}
-	return newLinks;
 }
+
